refactor: removed copyright at top of root command and changed package name to forms
