+++ src/cmd/root.go
@@ -1,6 +1,4 @@
-/*
-Copyright © 2024 NAME HERE <EMAIL ADDRESS>
-*/
+
 package cmd
 
 import (
